refactor: realocate ui components to specific folder
